third_party/blink/web_tests/http/tests/security/mixedContent/insecure-css-image-with-reload-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-subframe-and-reload.html' attempted to load subresource at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-image.html' over broken HTTPS.
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-image.html' attempted to load subresource at 'https://127.0.0.1:8443/security/mixedContent/resources/fetch-insecure-css-image.css' over broken HTTPS.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-subframe-and-reload.html' was loaded over HTTPS, but requested an insecure image 'http://example.test:8080/security/resources/compass.jpg'. This content should also be served over HTTPS.
CONSOLE MESSAGE: line 12: Reloading...
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-subframe-and-reload.html' attempted to load subresource at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-image.html' over broken HTTPS.
CONSOLE ERROR: line 2: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-insecure-css-image.html' attempted to load subresource at 'https://127.0.0.1:8443/security/mixedContent/resources/fetch-insecure-css-image.css' over broken HTTPS.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/mixedContent/resources/frame-with-subframe-and-reload.html' was loaded over HTTPS, but requested an insecure image 'http://example.test:8080/security/resources/compass.jpg'. This content should also be served over HTTPS.
CONSOLE MESSAGE: line 17: Second load finished.
This test opens a window that loads a frame with a subframe which contains insecure css image. The subframe is refreshed after it loads. We should trigger a mixed content callback twice despite the fact that on a second load of the subframe a css cache might be used to fetch an image.
